Move Claim flat file.
