+++ src/Libraries/Infrastructure/Confie.Infrastructure.Tests/FileHandling/FixedFileHandlingTests.cs
@@ -1,9 +1,12 @@
 ﻿using System;
 using System.IO;
 using Confie.Infrastructure.FileHandling;
-using Confie.WesternGeneral.FlatFile;
+using Confie.WesternGeneral;
 using NUnit.Framework;
 using Shouldly;
+using Feature = Confie.WesternGeneral.FlatFile.Feature;
+using PaymentTransaction = Confie.WesternGeneral.FlatFile.PaymentTransaction;
+using ReserveTransaction = Confie.WesternGeneral.FlatFile.ReserveTransaction;
 
 namespace Confie.Infrastructure.Tests.FileHandling
 {
